Removed getToken() request during initialization and add tests for bad credentials
(1) No longer call getToken() at initialization time and defer to
individual services to handle Credential implementations which generate
invalid access tokens.
(2) Assume null access tokens are invalid since there is no use case
for using the Admin SDK with null access tokens. Unauthenticated access
is possible via the databaseAuthVariableOverride parameter.
(3) Add tests for (a) apps which returns null access tokens, (b) apps
which return malformed access tokens, and (c) apps which fail to
generate access tokens. I added these tests for all Auth and FCM APIs.
Testing the Database is doable, but going to be really tough. Not only
can it not be initialized as easy as the other services, but it doesn’t
actually fulfill promises in these error conditions and instead just
logs to the console. I am going to punt on these tests for now and
continue to manually test these situations.

Note that verifyIdToken() and createCustomToken() do not use access
tokens and therefore can be used even if the app has an invalid
Credential.

Change-Id: I6330d17771b19dc752ef60b9be298890f8896a13